+++ src/test/groovy/nebula/test/ConcreteIntegrationSpec.groovy
@@ -1,15 +1,33 @@
 package nebula.test
 
-import org.gradle.BuildResult
+import nebula.test.functional.ExecutionResult
+import nebula.test.functional.internal.launcherapi.LauncherExecutionResult
+import org.gradle.api.logging.LogLevel
+import spock.lang.Unroll
 
 class ConcreteIntegrationSpec extends IntegrationSpec {
     def 'runs build'() {
         when:
-        BuildResult buildResult = runTasks('dependencies')
+        ExecutionResult buildResult = runTasks('dependencies')
+
+        then:
+        useToolingApi
+        buildResult.failure == null
+    }
+
+    def 'runs build with Launcher'() {
+        when:
+        useToolingApi = false
+        logLevel = LogLevel.DEBUG
+        ExecutionResult buildResult = runTasks('dependencies')
 
         then:
         buildResult.failure == null
-        buildResult.gradle != null
+        buildResult instanceof LauncherExecutionResult
+        ((LauncherExecutionResult) buildResult).gradle != null
+
+        cleanup:
+        useToolingApi = true
     }
 
     def 'setup and run build'() {
@@ -24,10 +42,32 @@ class ConcreteIntegrationSpec extends IntegrationSpec {
         fileExists('src/main/java/nebula/test/hello/HelloWorld.java')
 
         when:
-        runTasksSuccessfully('build')
+        def result = runTasksSuccessfully('build')
 
         then:
         fileExists('build/classes/main/nebula/test/hello/HelloWorld.class')
-        getStandardOutput().contains(':compileTestJava')
+        result.getStandardOutput().contains(':compileTestJava')
+    }
+
+
+    @Unroll
+    def 'can import from classpath using #desc #testTooling'(String desc, boolean testTooling) {
+        useToolingApi = testTooling
+
+        buildFile << '''
+            import nebula.test.FakePlugin
+            apply plugin: FakePlugin
+        '''.stripIndent()
+
+        when:
+        runTasksSuccessfully('tasks')
+
+        then:
+        noExceptionThrown()
+
+        where:
+        desc       | testTooling
+        "Tooling"  | true
+        "Launcher" | false
     }
 }